// src/types/generation/constants.ts
export enum GENERATION_REQUESTS {
    TEXT2IMAGE = "TEXT2IMAGE",
    IMAGE2IMAGE = 'IMAGE2IMAGE',
    ANIMATE = 'ANIMATE',
    INPAINT = 'INPAINT',
    REMOVE_BG = 'REMOVE_BG',
    TEXT2VIDEO = 'TEXT2VIDEO',
    TEXT2MUSIC = 'TEXT2MUSIC',
    FOLEY = 'FOLEY',
    TEXT2SPEECH = 'TEXT2SPEECH'
}

export const CHUBBY_KITTY = "https://avatars.charhub.io/avatars/uploads/images/gallery/file/1b9649dd-69f7-4231-802f-22be084ad0f9/c5a93b6a-7c28-45c0-bbdd-62b45d73bfdc.png";